Use doubles instead of floats

+++ src/main/scala/lattice/ascheme/ASchemeLatticeImpl.scala
@@ -59,7 +59,7 @@ class MakeASchemeLattice[LSeq : IsSchemeLattice] extends ASchemeLattice {
 
     def inject(x: Int) = Value(seq = lat.inject(x))
     def intTop = Value(seq = lat.intTop)
-    def inject(x: Float) = Value(seq = lat.inject(x))
+    def inject(x: Double) = Value(seq = lat.inject(x))
     def inject(x: String) = Value(seq = lat.inject(x))
     def inject(x: Boolean) = Value(seq = lat.inject(x))
     def inject(x: Char) = Value(seq = lat.inject(x))
